db: own the vector index artifact catalogue
