+++ adapters/repos/db/vector_index_artifacts_test.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package helpers
+package db
 
 import (
 	"regexp"
